memperbaiki view.php

+++ view.php
@@ -20,9 +20,12 @@ $result = mysqli_query($konek, $sql);
             <div class="col m-auto">
                 <div class="card mt-5 p-3 border-0">
                     <h4 class="p-3">Records Regristasi</h4>
+                    <a href="" class="btn btn-sm btn-success">Add new record</a>
+
                     <table class="table table-bordered">
                         <thead>
                             <tr>
+                                <th scope="col">No</th>
                                 <th scope="col">User ID</th>
                                 <th scope="col">Username</th>
                                 <th scope="col">Email</th>
@@ -32,25 +35,33 @@ $result = mysqli_query($konek, $sql);
                         </thead>
                         <tbody>
                             <?php
+                            $i = 1;
+                            if (mysqli_num_rows($result) > 0)
+                            {
                             while ($row = mysqli_fetch_assoc($result)) {
                                 $userid = $row['user_id'];
                                 $username = $row['username'];
                                 $email = $row['email'];
                                 $age = $row['age'];
-                            }
                             ?>
                                 <tr>
-                                    <th scope="row"><?php echo $userid ?></th>
+                                    <th scope="row"><?php echo $i++ ?></th>
+                                    <th><?php echo $userid ?></th>
                                     <td><?php echo $username ?></td>
                                     <td><?php echo $email ?></td>
                                     <td><?php echo $age ?></td>
                                     <td>
                                         <a href="edit.php?id=<?php echo $userid ?>" class="btn btn-sm btn-success">Edit</a>
-                                        <a href="delete.php?id=<?php echo $userid ?>" class="btn btn-sm btn-danger">Delete</a>
+                                        <a href="delete.php?id=<?php echo $userid ?>" onclick="return confirm('Apakah anda yakin ingin menghapusnya?')" class="btn btn-sm btn-danger">Delete</a>
                                     </td>
                                 </tr>
+                                <?php }
+
+                                } else $kosong = "<h4 class='d-flex justify-content-center text-danger'>Tidak ada records</h4>";
+                                ?>
                         </tbody>
                     </table>
+                    <?php echo $kosong ?>
                 </div>
             </div>
         </div>
